requirement.txt
# Google Generative AI API
google-generativeai==0.2.0

# For handling PDF files
PyPDF2==3.0.1

# For handling DOCX files
python-docx==0.8.11

# General purpose libraries
PyYAML==6.0  # For reading config.yaml
